+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # v1.10.0
 ## mm/dd/mmmm
 
+1. [](#improved)
+    * Various notifications improvements
 1. [](#bugfix)
     * Fixed missed highlight on the selected page in Parents field
+    * Fixed notifications that would not be remembered as hidden
 
 # v1.10.0-rc.20
 ## 12/14/2020